// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-workspace-initial-configuration-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingDescriptionBody } from '../models/linking-description-body';
import { LinkingSelectedWordBody } from '../models/linking-selected-word-body';
export interface LinkingWorkspaceInitialConfigurationBody {
  automaticWordMatchesEnabled: boolean | null;
  ayahOverrides: Array<number> | null;
  descriptions: Array<LinkingDescriptionBody> | null;
  inclusionMode: string | null;
  manualLinkShape: string | null;
  selectedWords: Array<LinkingSelectedWordBody> | null;
}