// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-workspace-source-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingSelectedWordBody } from '../models/linking-selected-word-body';
import { LinkingSourceDescriptorBody } from '../models/linking-source-descriptor-body';
import { LinkingWorkspaceDescriptionResponse } from '../models/linking-workspace-description-response';
import { LinkingWorkspaceManualAyahResponse } from '../models/linking-workspace-manual-ayah-response';
export interface LinkingWorkspaceSourceResponse {
  automaticWordMatchesEnabled: boolean | null;
  ayahOverrides: Array<number>;
  descriptions: Array<LinkingWorkspaceDescriptionResponse>;
  descriptor: LinkingSourceDescriptorBody;
  id: number;
  inclusionMode: string;
  lastResolvedAtUtc: string | null;
  lastResolvedCount: number | null;
  manualAyahs: Array<LinkingWorkspaceManualAyahResponse>;
  manualLinkShape: string | null;
  orderValue: number;
  selectedWords: Array<LinkingSelectedWordBody>;
  sourceIdentity: string;
  sourceVersion: number;
}